perf(cli): reduce memory footprint of diagnostic printer

+++ crates/biome_lsp/src/session.rs
@@ -25,6 +25,7 @@ use std::sync::atomic::AtomicU8;
 use std::sync::atomic::Ordering;
 use std::sync::Arc;
 use std::sync::RwLock;
+use std::u32;
 use tokio::sync::Notify;
 use tokio::sync::OnceCell;
 use tower_lsp::lsp_types;
@@ -310,7 +311,7 @@ impl Session {
             let result = self.workspace.pull_diagnostics(PullDiagnosticsParams {
                 path: rome_path,
                 categories,
-                max_diagnostics: u64::MAX,
+                max_diagnostics: u32::MAX,
             })?;
 
             tracing::trace!("biome diagnostics: {:#?}", result.diagnostics);
